Add support for packs and internal engines
When using packs or internal engines, the template paths may not be
nested directly under the Rails root `app/views` directory.

For example, the default for packs-rails is to use a `packs` directory
under the Rails root, so the layout could be in a directory like
`#{Rails.root}/packs/pack_name/app/views`.

Since the view paths are registere with
`ActionController::Base.view_paths`, we can use that collection to
clearly identify local templates